Attempt to fix bug where reading from the opened socket where randomly not possible
It looks like after the firts disconnection, waiting for any the other
sockets to be ready for reading was randomly timeout'ing. When
disconnecting, the corresponding socket wasn't properly closed and
removed from the select'ed list of filehandles. As the bug was quite
random, I suspect these two missing close were leading unespected
behaviours.