Fix parsers to cleans bytes quickly
We were waiting too long to clean unknown bytes from the buffer.
Previously, the v3 parser was waiting for more bits from the backend if
it can not understand them. We can drop these packets immediately as
soon as they are bigger than 5 bytes (a minimal message length) or 1
bytes for SSL dialog.