Change session tracking structure format
We used to track sessions TCP dialog using an array: 0 for frontend data
and properties, 1 for backend ones. This patch change this to use a hash
instead of this array using keys 'F' and 'B' for each side of the
communication. This will help to track properties globally to the
session.