Teach pgShark how to detect the backend
If the 'host' parameter is not given when creating a pgShark object,
then it will not call user specified callback unless it can identify
with no doubt who is the backend.
This is based on function get_msg_type_*. As soon as one is returning a
type and NOT the other one, then we can identify one side of the dialog
and deduce who is the backend.